# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * clipp'n'k API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Me {
  id: number;
  discordId: string;
  username: string;
  /** @nullable */
  avatarUrl: string | null;
  isAdmin: boolean;
  usedStorageBytes: number;
  quotaStorageBytes: number;
  /** Effective per-file upload size limit for this user (role-aware), used by the upload dialog to filter oversized files client-side. */
  maxUploadBytes: number;
  /** Discord role IDs this user currently holds in the configured guild. */
  roles: string[];
  /** Whether an admin has banned this account. */
  banned: boolean;
  /**
     * GitHub login of the linked account, null if not connected.
     * @nullable
     */
  githubUsername: string | null;
  /** Whether the +1 GB star bonus has already been permanently granted. */
  githubStarBonusGranted: boolean;
}

export interface GithubStarStatus {
  githubUsername: string;
  starred: boolean;
  bonusGranted: boolean;
}

export type ClipVisibility = typeof ClipVisibility[keyof typeof ClipVisibility];


export const ClipVisibility = {
  public: 'public',
  private: 'private',
} as const;

export type ClipStatus = typeof ClipStatus[keyof typeof ClipStatus];


export const ClipStatus = {
  processing: 'processing',
  ready: 'ready',
  failed: 'failed',
} as const;

export interface Clip {
  id: number;
  slug: string;
  title: string;
  originalFilename: string;
  mimeType: string;
  sizeBytes: number;
  /** @nullable */
  durationSeconds: number | null;
  /** @nullable */
  width: number | null;
  /** @nullable */
  height: number | null;
  /**
     * Owner-authenticated playback URL, null while the clip is still processing. Distinct from PublicClip.videoUrl, which is only served for clips with visibility=public.
     * @nullable
     */
  videoUrl: string | null;
  /** @nullable */
  thumbnailUrl: string | null;
  visibility: ClipVisibility;
  status: ClipStatus;
  /** @nullable */
  failureReason?: string | null;
  shareUrl: string;
  createdAt: string;
  updatedAt: string;
}

/**
 * Multipart form upload. The `file` field carries raw video bytes; the frontend performs this upload with a hand-rolled fetch/XHR call (for upload-progress reporting) rather than the generated hook.
 */
export interface ClipUploadInput {
  file: string;
  title?: string;
}

export interface ClipUpdate {
  /** @minLength 1 */
  title?: string;
  visibility?: ClipVisibility;
}

export interface ClipTrimInput {
  /** @minimum 0 */
  startSeconds: number;
  /** @minimum 0 */
  endSeconds: number;
  /**
     * @minimum 0
     * @nullable
     */
  cropX?: number | null;
  /**
     * @minimum 0
     * @nullable
     */
  cropY?: number | null;
  /**
     * @minimum 1
     * @nullable
     */
  cropWidth?: number | null;
  /**
     * @minimum 1
     * @nullable
     */
  cropHeight?: number | null;
}

export interface PublicClip {
  slug: string;
  title: string;
  videoUrl: string;
  /** @nullable */
  thumbnailUrl: string | null;
  /** @nullable */
  width: number | null;
  /** @nullable */
  height: number | null;
  /** @nullable */
  durationSeconds: number | null;
  ownerUsername: string;
}

export interface PublicSiteSettings {
  brandingTitle: string;
  /** @nullable */
  brandingLogoUrl: string | null;
  brandingPrimaryColor: string;
}

export interface SiteSettings {
  /** @nullable */
  discordGuildId: string | null;
  /**
     * Discord channel ID where "Share to Discord" posts are sent, null if disabled.
     * @nullable
     */
  discordShareChannelId: string | null;
  brandingTitle: string;
  /** @nullable */
  brandingLogoUrl: string | null;
  brandingPrimaryColor: string;
  maxUploadBytes: number;
  maxUserStorageBytes: number;
  /** @nullable */
  maxClipDurationSeconds: number | null;
  allowedMimeTypes: string[];
  defaultVisibility: ClipVisibility;
  /** True when GITHUB_CLIENT_ID, GITHUB_CLIENT_SECRET and GITHUB_REDIRECT_URI are all set in the server environment (i.e. the star-bonus flow is functional). */
  githubBonusEnabled: boolean;
  /** True when DISCORD_CLIENT_ID, DISCORD_CLIENT_SECRET and DISCORD_REDIRECT_URI are all set in the server environment (i.e. Discord OAuth login is functional). */
  discordEnabled: boolean;
  /** True when DISCORD_BOT_TOKEN is set in the server environment (i.e. guild membership verification is functional). */
  discordBotEnabled: boolean;
}

export interface SiteSettingsUpdate {
  /** @nullable */
  discordGuildId?: string | null;
  /** @nullable */
  discordShareChannelId?: string | null;
  /** @minLength 1 */
  brandingTitle?: string;
  /** @nullable */
  brandingLogoUrl?: string | null;
  brandingPrimaryColor?: string;
  /** @minimum 1 */
  maxUploadBytes?: number;
  /** @minimum 1 */
  maxUserStorageBytes?: number;
  /**
     * @minimum 1
     * @nullable
     */
  maxClipDurationSeconds?: number | null;
  allowedMimeTypes?: string[];
  defaultVisibility?: ClipVisibility;
}

export interface AdminUser {
  id: number;
  discordId: string;
  username: string;
  /** @nullable */
  avatarUrl: string | null;
  usedStorageBytes: number;
  quotaStorageBytes: number;
  /**
     * Admin-set storage quota override, null when role/site defaults apply.
     * @nullable
     */
  quotaOverrideBytes: number | null;
  roles: string[];
  banned: boolean;
  isAdmin: boolean;
  clipCount: number;
  createdAt: string;
}

export interface AdminUserUpdate {
  banned?: boolean;
  /** @nullable */
  quotaOverrideBytes?: number | null;
}

export interface DiscordGuildRole {
  id: string;
  name: string;
  position: number;
}

export interface DiscordRoleLimit {
  roleId: string;
  roleName: string;
  /** Higher wins when a user holds multiple configured roles. */
  priority: number;
  /**
     * Per-file upload limit; null inherits the site default.
     * @nullable
     */
  maxUploadBytes: number | null;
  /**
     * Storage quota; null inherits the site default.
     * @nullable
     */
  maxUserStorageBytes: number | null;
}