# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: clipp'n'k API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /clips:
    get:
      operationId: listClips
      tags: [clips]
      summary: List the current user's clips
      responses:
        "200":
          description: Clip list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Clip"
        "401":
          description: Not logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: uploadClip
      tags: [clips]
      summary: Upload a new clip (multipart form upload)
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/ClipUploadInput"
      responses:
        "201":
          description: Clip created and queued for processing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Clip"
        "400":
          description: Invalid file
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "413":
          description: File too large or quota exceeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /admin/settings:
    get:
      operationId: getAdminSettings
      tags: [admin]
      summary: Get site settings (admin only)
      responses:
        "200":
          description: Site settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SiteSettings"
        "403":
          description: Not an admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateAdminSettings
      tags: [admin]
      summary: Update site settings (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SiteSettingsUpdate"
      responses:
        "200":
          description: Updated site settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SiteSettings"
        "403":
          description: Not an admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/users:
    get:
      operationId: listAdminUsers
      tags: [admin]
      summary: List all users with usage and status (admin only)
      responses:
        "200":
          description: User list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminUser"
        "403":
          description: Not an admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/users/{id}:
    patch:
      operationId: updateAdminUser
      tags: [admin]
      summary: Ban/unban a user or set their storage quota override (admin only)
      parameters:
        - $ref: "#/components/parameters/UserId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteAdminUser
      tags: [admin]
      summary: Delete a user and all of their clips (admin only)
      parameters:
        - $ref: "#/components/parameters/UserId"
      responses:
        "204":
          description: User deleted
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/discord/guild-roles:
    get:
      operationId: listDiscordGuildRoles
      tags: [admin]
      summary: Fetch the roles of the configured Discord guild (admin only)
      responses:
        "200":
          description: Guild roles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DiscordGuildRole"
        "400":
          description: No guild configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/discord/role-limits:
    get:
      operationId: listDiscordRoleLimits
      tags: [admin]
      summary: List configured per-role upload/storage limits (admin only)
      responses:
        "200":
          description: Role limit configuration
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DiscordRoleLimit"
    put:
      operationId: updateDiscordRoleLimits
      tags: [admin]
      summary: Replace the per-role upload/storage limit configuration (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: "#/components/schemas/DiscordRoleLimit"
      responses:
        "200":
          description: Saved role limit configuration
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DiscordRoleLimit"
components:
  parameters:
...
    UserId:
      name: id
      in: path
      required: true
      schema:
        type: integer
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    ClipVisibility:
      type: string
      enum: [public, private]
    ClipStatus:
      type: string
      enum: [processing, ready, failed]
    Clip:
      type: object
      required:
        [
          id,
          slug,
          title,
          originalFilename,
          mimeType,
          sizeBytes,
          durationSeconds,
          width,
          height,
          videoUrl,
          thumbnailUrl,
          visibility,
          status,
          shareUrl,
          createdAt,
          updatedAt,
        ]
      properties:
        id: { type: integer }
        slug: { type: string }
        title: { type: string }
        originalFilename: { type: string }
        mimeType: { type: string }
        sizeBytes: { type: integer }
        durationSeconds: { type: ["number", "null"] }
        width: { type: ["integer", "null"] }
        height: { type: ["integer", "null"] }
        videoUrl:
          type: ["string", "null"]
          description: >-
            Owner-authenticated playback URL, null while the clip is still
            processing. Distinct from PublicClip.videoUrl, which is only
            served for clips with visibility=public.
        thumbnailUrl: { type: ["string", "null"] }
        visibility: { $ref: "#/components/schemas/ClipVisibility" }
        status: { $ref: "#/components/schemas/ClipStatus" }
        failureReason: { type: ["string", "null"] }
        shareUrl: { type: string }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    ClipUploadInput:
      type: object
      required: [file]
      description: >-
        Multipart form upload. The `file` field carries raw video bytes; the
        frontend performs this upload with a hand-rolled fetch/XHR call (for
        upload-progress reporting) rather than the generated hook.
      properties:
        file:
          type: string
        title:
          type: string
...
    SiteSettings:
      type: object
      required:
        [
          discordGuildId,
          discordShareChannelId,
          brandingTitle,
          brandingLogoUrl,
          brandingPrimaryColor,
          maxUploadBytes,
          maxUserStorageBytes,
          maxClipDurationSeconds,
          allowedMimeTypes,
          defaultVisibility,
          githubBonusEnabled,
          discordEnabled,
          discordBotEnabled,
        ]
      properties:
        discordGuildId: { type: ["string", "null"] }
        discordShareChannelId:
          type: ["string", "null"]
          description: Discord channel ID where "Share to Discord" posts are sent, null if disabled.
        brandingTitle: { type: string }
        brandingLogoUrl: { type: ["string", "null"] }
        brandingPrimaryColor: { type: string }
        maxUploadBytes: { type: integer }
        maxUserStorageBytes: { type: integer }
        maxClipDurationSeconds: { type: ["integer", "null"] }
        allowedMimeTypes:
          type: array
          items: { type: string }
        defaultVisibility: { $ref: "#/components/schemas/ClipVisibility" }
        githubBonusEnabled:
          type: boolean
          description: >-
            True when GITHUB_CLIENT_ID, GITHUB_CLIENT_SECRET and GITHUB_REDIRECT_URI
            are all set in the server environment (i.e. the star-bonus flow is functional).
        discordEnabled:
          type: boolean
          description: >-
            True when DISCORD_CLIENT_ID, DISCORD_CLIENT_SECRET and DISCORD_REDIRECT_URI
            are all set in the server environment (i.e. Discord OAuth login is functional).
        discordBotEnabled:
          type: boolean
          description: >-
            True when DISCORD_BOT_TOKEN is set in the server environment
            (i.e. guild membership verification is functional).
    SiteSettingsUpdate:
      type: object
      properties:
        discordGuildId: { type: ["string", "null"] }
        discordShareChannelId: { type: ["string", "null"] }
        brandingTitle: { type: string, minLength: 1 }
        brandingLogoUrl: { type: ["string", "null"] }
        brandingPrimaryColor: { type: string }
        maxUploadBytes: { type: integer, minimum: 1 }
        maxUserStorageBytes: { type: integer, minimum: 1 }
        maxClipDurationSeconds: { type: ["integer", "null"], minimum: 1 }
        allowedMimeTypes:
          type: array
          items: { type: string }
        defaultVisibility: { $ref: "#/components/schemas/ClipVisibility" }
    AdminUser:
      type: object
      required:
        [
          id,
          discordId,
          username,
          avatarUrl,
          usedStorageBytes,
          quotaStorageBytes,
          quotaOverrideBytes,
          roles,
          banned,
          isAdmin,
          clipCount,
          createdAt,
        ]
      properties:
        id: { type: integer }
        discordId: { type: string }
        username: { type: string }
        avatarUrl: { type: ["string", "null"] }
        usedStorageBytes: { type: integer }
        quotaStorageBytes: { type: integer }
        quotaOverrideBytes:
          type: ["integer", "null"]
          description: Admin-set storage quota override, null when role/site defaults apply.
        roles:
          type: array
          items: { type: string }
        banned: { type: boolean }
        isAdmin: { type: boolean }
        clipCount: { type: integer }
        createdAt: { type: string, format: date-time }
    AdminUserUpdate:
      type: object
      properties:
        banned: { type: boolean }
        quotaOverrideBytes: { type: ["integer", "null"] }
    DiscordGuildRole:
      type: object
      required: [id, name, position]
      properties:
        id: { type: string }
        name: { type: string }
        position: { type: integer }
    DiscordRoleLimit:
      type: object
      required: [roleId, roleName, priority, maxUploadBytes, maxUserStorageBytes]
      properties:
        roleId: { type: string }
        roleName: { type: string }
        priority:
          type: integer
          description: Higher wins when a user holds multiple configured roles.
        maxUploadBytes:
          type: ["integer", "null"]
          description: Per-file upload limit; null inherits the site default.
        maxUserStorageBytes:
          type: ["integer", "null"]
          description: Storage quota; null inherits the site default.